\docType{package}
\name{tictoc}
\alias{tictoc}
\alias{tictoc-package}
\title{Package tictoc.}
\description{
Functions for timing, as well as implementations of Stack
and List structures.
}
\details{
This package provides the timing functions \code{tic} and
\code{toc} that can be nested. One can record all timings
while a complex script is running, and examine the values
later.

In addition, this package provides classes
(\code{\link{Stack}}), implemented as a \code{vector}, and
\code{\link{List}}, implemented as a \code{list}, both of
which support operations \code{push}, \code{pop},
\code{first}, \code{last}, \code{clear} and \code{size}.
}
\section{Copyright}{
  Copyright (C) Collective, Inc.
}

\section{License}{
  Apache License, Version 2.0, available at
  http://www.apache.org/licenses/LICENSE-2.0
}

\section{URL}{
  http://github.com/collectivemedia/tictoc
}
\author{
Sergei Izrailev
}
\keyword{list}
\keyword{profiling}
\keyword{stack}
\keyword{timing}

